Latest and Greatest Grapher
This Grapher has a new file called "SensorKeneticsGrapher.py".  This one
has a user iterface identical to "BlackBoxGrapher.py", but works with
the format outputted by the Sensor Kinetics app.  If I need to create a
Grapher for another Grapher, I'll take the time to make the format
adjustable, the data fluid, and the actual Graph maker to accept funkier
data.